ENH: initial work on image library for PI-RADS report
The only difference from the standard Image library is the presence of the
acquisition type code at the image library group level, which identifies
the type of sequence that corresponds to the items stored in the group. It is
implied that each individual series should be communicated in a separate
image library group.